Refactor ChatClient class
