+++ stream-chat-android-client/src/test/java/io/getstream/chat/android/client/ChatClientImplTest.kt
@@ -53,7 +53,7 @@ internal class ChatClientImplTest {
         )
 
         socket = FakeChatSocket()
-        client = ChatClientImpl(
+        client = ChatClient(
             config = config,
             api = mock(),
             socket = socket,
