+++ stream-chat-android-client/src/test/java/io/getstream/chat/android/client/MockClientBuilder.kt
@@ -77,7 +77,7 @@ internal class MockClientBuilder {
             fileUploader
         )
 
-        client = ChatClientImpl(config, api, socket, notificationsManager)
+        client = ChatClient(config, api, socket, notificationsManager)
         client.setUser(user, token)
 
         socket.sendEvent(connectedEvent)
